Exception handling - delete

+++ src/main/java/com/andrezktt/project/services/UserService.java
@@ -2,8 +2,11 @@ package com.andrezktt.project.services;
 
 import com.andrezktt.project.entities.User;
 import com.andrezktt.project.repositories.UserRepository;
+import com.andrezktt.project.services.exceptions.DatabaseException;
 import com.andrezktt.project.services.exceptions.ResourceNotFoundException;
+import org.hibernate.exception.DataException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +32,15 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        try {
+            if (repository.existsById(id)) {
+                repository.deleteById(id);
+            } else {
+                throw new ResourceNotFoundException(id);
+            }
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
